Finish Project

+++ app/Http/Controllers/CarController.php
@@ -4,7 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Car;
 use App\Location;
+use App\User;
+use App\Payment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Auth;
 
 class CarController extends Controller
 {
@@ -23,6 +27,7 @@ class CarController extends Controller
         return view ('cars.index')->with('cars', $cars);
     }
 
+
     /**
      * Show the form for creating a new resource.
      *
@@ -30,6 +35,10 @@ class CarController extends Controller
      */
     public function create(Location $location)
     {
+        if(Gate::denies('cars-management')){
+            return redirect(route('cars'));
+        }
+
         return view ('cars.create')->with('location', $location);
     }
 
@@ -59,6 +68,12 @@ class CarController extends Controller
         return view ('cars.show', compact('car', 'location'));
     }
 
+    public function showCarsFromOwner()
+    {
+        $cars = Car::with('user','location')->where('user_id', '=', Auth::id() )->get();
+        return view ('cars.mycars')->with('cars', $cars);
+    }
+
     /**
      * Show the form for editing the specified resource.
      *
@@ -67,6 +82,10 @@ class CarController extends Controller
      */
     public function edit(Car $car)
     {
+        if(Gate::denies('cars-management')){
+            return redirect(route('cars'));
+        }
+
         return view ('cars.edit', compact('car'));
     }
 
@@ -79,6 +98,11 @@ class CarController extends Controller
      */
     public function update(Request $request, Car $car)
     {
+        if($request->status){
+
+            $car->update(['status' => $request->status]);
+        }
+
         $car->update($this->validateLocations($request));
 
         return redirect('cars')->with('success', 'Car updated!');
@@ -92,6 +116,10 @@ class CarController extends Controller
      */
     public function destroy(Car $car)
     {
+        if(Gate::denies('cars-management')){
+            return redirect(route('cars'));
+        }
+
         $car->delete();
         return redirect('cars')->with('success', 'Successfully deleted car!');
     }
